authority-server/requirements.txt
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.6
dotenv==0.9.9
flask==3.1.3
flask-cors==6.0.2
flask-socketio==5.6.1
flask-sqlalchemy==3.1.1
greenlet==3.2.5
h11==0.16.0
importlib-metadata==8.7.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
pycparser==2.23
pyjwt==2.12.1
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
sqlalchemy==2.0.48
typing-extensions==4.15.0
werkzeug==3.1.7
wsproto==1.2.0
zipp==3.23.0
